test(types): move sequence test to combinators, start combinators testing
